backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.24.3
